handled profile edit for own profile page
